refactor(tech): delete dependent types - delete DBBase dependent types bcz are not provide from domain - transfer to app/interface
