+++ src/modules/tech/application/tech-read.usecase.ts
@@ -2,9 +2,8 @@ import { Inject, Injectable } from '@nestjs/common';
 import {
   FullTechData,
   LengBase,
-  ReadAllFlattenTechsRes,
 } from 'src/domain/entities/tech';
-import { TechRepository } from './tech.interface';
+import { ReadAllFlattenTechsRes, TechRepository } from './tech.interface';
 import { TECH_REPOSITORY } from 'src/modules/tokens';
 import { DBBase } from '@/dynamic.types';
