+++ src/modules/tech/application/tech.interface.ts
@@ -3,8 +3,21 @@
 import { DBBase } from '@/dynamic.types';
 import { CRUI } from '@/shareds/pattern/application/interfaces/cru';
 import { DeleteProps } from '@/shareds/pattern/application/interfaces/delete';
-import { Leng, LengBase, TechBase, TechForm } from 'src/domain/entities/tech';
-import { MongooseBase } from 'src/shareds/pattern/infrastructure/types/mongoose';
+import { FullTechData, LengBase, LibBase, TechBase, TechForm } from 'src/domain/entities/tech';
+
+export type Fw = TechBase & DBBase & {
+  librerias?: (LibBase&DBBase)[];
+}
+export type Leng = TechBase & DBBase & {
+  frameworks?: Fw[];
+}
+
+export type ReadAllFlattenTechsRes = {
+  techs: (LengBase & DBBase)[];
+  flattenTechs: FullTechData[];
+  dispoFw: { name: string }[];
+  dispoLeng: { name: string }[];
+};
 
 export interface TechRepository extends CRUI<LengBase> {
   read(filter?: Partial<LengBase & DBBase>): Promise<Array<Leng>>;
