+++ src/shareds/pattern/infrastructure/types/mongoose.d.ts
@@ -1,10 +1,5 @@
 import mongoose, { Document, QueryOptions, UpdateQuery } from 'mongoose';
 
-export type MongooseBase = {
-  id: string;
-  createdAt: string;
-  updatedAt: string;
-};
 export type MongooseTimestamps = {
   createdAt: Date;
   updatedAt: Date;
